refactor and make pearlclip scanning better probably
Make ServerTeleport and OnUpdateWalkingPlayerPre more abstracted; refactor Secret Aura; remove removeCameraInterpolation and replace it with a param in setPlayerPosition; change behavior of floor2D and floor3D in Vector3 and remove the static methods to create one in favor of putting it in the constructor itself. I probably fucked something.